Flesh out options examples with title, description, etc
https://github.com/sinclairzx81/typebox/issues/9#issuecomment-894818860

I had searched for "title", "description", "document", and "nullable" in search of how to add annotations.  
Many JSON Schema users use it for OpenAPI, which has the `nullable` property that isn't in JSON Schema, and may be curious how to use it. They may also wish to know whether extensions like `x-mycompany-myattribute` work.